requirements.txt
streamlit
google-generativeai
python-dotenv
pdfminer.six
pandas
python-docx